requirements.txt
imutils
scikit-learn==0.24.2
opencv-python
numpy==1.19.5
tensorflow==2.5.0
Keras==2.4.3
tqdm==4.61.1
PyYAML